requirements.txt
Django~=5.1.4
djangorestframework~=3.15.2
django-debug-toolbar~=4.2.0
pillow ~=11.1.0
django-cors-headers~=4.7.0
django-oauth-toolkit~=3.0.1
